supabase 연동 완료 쿼리 수정 fetch 오류에 관한 확인 처리 추가

+++ backend/src/main/java/com/kdigital/miniproject/persistence/FetchLogRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kdigital.miniproject.domain.FetchLog;
 
-public interface FetchLogRepository extends JpaRepository<FetchLog, Integer> {
+public interface FetchLogRepository extends JpaRepository<FetchLog, Long> {
 	Page<FetchLog> findByErrorMsgIsNull(Pageable pageable);
 	Page<FetchLog> findByErrorMsgIsNotNull(Pageable pageable);
 }
